backend/requirements.txt
# Backend dependencies
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
torch==2.1.0
transformers==4.34.0
datasets==2.14.5
matplotlib==3.8.0
seaborn==0.13.0
pytest==7.4.2
pytest-flask==1.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
